[AOT Refactor] graph signature to i/o analysis
[ghstack-poisoned]
