+++ torch/_functorch/aot_autograd.py
@@ -66,7 +66,6 @@ from ._aot_autograd.schemas import (
     MutationType,
     OutputType,
     GraphSignature,
-    BackwardSignature,
     AOTConfig,
     SubclassTracingInfo,
 )
@@ -84,6 +83,7 @@ from ._aot_autograd.input_output_analysis import (
     remove_dupe_metadata,
     merge_view_inputs,
     create_synthetic_base_metadata,
+    create_graph_signature,
 )
 from ._aot_autograd.runtime_wrappers import (
     create_runtime_wrapper,
@@ -2143,88 +2143,6 @@ def create_functional_call(mod, params_spec, params_len):
         return out
     return functional_call
 
-
-def _graph_input_names(gm):
-    return [node.name for node in gm.graph.nodes if node.op == "placeholder"]
-
-
-def _graph_output_names(gm):
-    output_node = next(iter(reversed(gm.graph.nodes)))
-    assert output_node.op == "output" and len(output_node.args) == 1
-    return_args = output_node.args[0]
-    return [getattr(return_arg, "name", None) for return_arg in return_args]
-
-
-def create_graph_signature(
-    fx_g: torch.fx.GraphModule,
-    fw_metadata: ViewAndMutationMeta,
-    in_spec: pytree.TreeSpec,
-    out_spec: pytree.TreeSpec,
-    *,
-    user_args_flat: List[torch.Tensor],
-    params_and_buffers_flat: List[torch.Tensor],
-    param_names: List[str],
-    buffer_names: List[str],
-    trace_joint: bool,
-    num_user_fw_outs: Optional[int],
-    loss_index: Optional[int],
-) -> GraphSignature:
-
-    # Retrieve graph input names
-    graph_input_names = _graph_input_names(fx_g)
-    # Retrieve graph output names
-    graph_output_names = _graph_output_names(fx_g)
-
-    num_params_buffers = len(param_names) + len(buffer_names)
-    # We have enough restrictions on the graph (no de-duping, synthetic bases, etc),
-    # Such that # graph inps = # user inps + # params + # buffers
-    num_user_args = len(graph_input_names) - num_params_buffers
-
-    if trace_joint:
-        assert num_user_fw_outs is not None
-        num_fw_outs = num_user_fw_outs + fw_metadata.num_mutated_inp_runtime_indices
-        backward_output_names = graph_output_names[num_fw_outs:]
-
-        grad_index = itertools.count(0)
-        gradients_to_parameters = {
-            backward_output_names[next(grad_index)]: param_names[i]
-            for i, param in enumerate(params_and_buffers_flat)
-            if param.requires_grad
-        }
-
-        gradients_to_user_inputs = {
-            backward_output_names[next(grad_index)]: graph_input_names[i + len(params_and_buffers_flat)]
-            for i, user_input in enumerate(user_args_flat)
-            if user_input.requires_grad
-        }
-
-        assert len(gradients_to_parameters) + len(gradients_to_user_inputs) == len(
-            backward_output_names
-        )
-
-        # Check that we have fully accounted for all graph outputs
-        backward_signature = BackwardSignature(
-            gradients_to_parameters,
-            gradients_to_user_inputs,
-            graph_output_names[loss_index],
-        )
-    else:
-        backward_signature = None
-        num_user_fw_outs = len(graph_output_names) - fw_metadata.num_mutated_inp_runtime_indices
-
-    return GraphSignature.from_tracing_metadata(
-        in_spec=in_spec,
-        out_spec=out_spec,
-        graph_input_names=graph_input_names,
-        graph_output_names=graph_output_names,
-        view_mutation_metadata=fw_metadata,
-        named_parameters=param_names,
-        named_buffers=buffer_names,
-        num_user_inputs=num_user_args,
-        num_user_outputs=num_user_fw_outs,
-        loss_index=loss_index,
-        backward_signature=backward_signature,
-    )
 
 def aot_function(
     fn: Callable,
